fix style: login/signup form The loginsignup component was moved to views folder
