+++ movies-and-series-front/src/components/Views/LoginSignup/LoginSignUp.css
@@ -19,12 +19,8 @@ form {
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  gap: 0.5rem;
+  gap: 0.7rem;
   width: 100%;
-}
-
-label {
-  font-size: small;
 }
 
 .container-title {
@@ -36,7 +32,7 @@ label {
 .inputs {
   border-radius: 0.3rem;
   border-width: 0.01rem;
-  height: 1.3rem;
+  height: 1.5rem;
 }
 
 .input-btn {
@@ -64,6 +60,20 @@ label {
   border: none;
 }
 
-.grey {
-  background: #858282;
+.toggle {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 5rem;
+  border: none;
+  background: rgb(47, 128, 128);
+  color: #d6d6d6;
+  border-radius: 0 0 10px 10px;
+  font-weight: 700;
 }
+
+.grey {
+  background: #d6d6d6;
+  color: rgb(0, 101, 101);
+}
+
